counter refactor for PDP
